refactor(frontend): split monolithic api.ts into per-domain modules (closes #295)
Split the 2372-line frontend/src/lib/api.ts into 15 per-domain modules
under frontend/src/lib/api/ with a barrel index.ts preserving the
@/lib/api import path. No behavioral changes.

Domain modules: core, tags, folders, sessions, groups, users,
vault-groups, wiki, kms, health, settings, vaults, organizations,
memories, auth-sessions

- 92 tests pass (6 test files)
- 0 circular dependencies (madge)
- SSE handler (parseSSEStream) unchanged
- typecheck, lint --max-warnings 0, and build all clean
