+++ frontend/src/lib/api/index.ts
@@ -1,0 +1,16 @@
+export * from "./core";
+export * from "./tags";
+export * from "./folders";
+export * from "./sessions";
+export * from "./groups";
+export * from "./users";
+export * from "./vault-groups";
+export * from "./wiki";
+export * from "./kms";
+export * from "./health";
+export * from "./settings";
+export * from "./vaults";
+export * from "./organizations";
+export * from "./memories";
+export * from "./auth-sessions";
+export { default } from "./core";
